Untangle the position where config is initialized
=============================== Coverage summary ================
Statements   : 95.76% ( 4067/4247 ), 6 ignored
Branches     : 92.79% ( 1776/1914 ), 39 ignored
Functions    : 95.82% ( 413/431 ), 6 ignored
Lines        : 96.42% ( 3501/3631 )
=================================================================

File dist/finitedomain.dist.min.js created: 193.2 kB → 51.86 kB → 15.53 kB (gzip)
